+++ src/features/expenses/expenses.swagger.api.ts
@@ -1,3 +1,5 @@
+import { swaggerErrorNotFound } from "../../routes/middlewares/swagger/swagger.errors";
+
 export const expensesSwagger = {
     "/api/expenses": {
         "get": {
@@ -10,7 +12,15 @@ export const expensesSwagger = {
                 "200": {
                     "description": "OK",
                     "content": {
-                        "application/json": {}
+                        "application/json": {
+                            "data": {
+                                "type": "array",
+                                "description": "Array of expense objects",
+                                "items": {
+                                    "$ref": "#/components/schemas/Expense"
+                                }
+                            }
+                        }
                     }
                 }
             }
@@ -89,22 +99,7 @@ export const expensesSwagger = {
                         }
                     }
                 },
-                "404": {
-                    "description": "Not Found",
-                    "content": {
-                        "application/json": {
-                            "schema": {
-                                "type": "object",
-                                "properties": {
-                                    "error": {
-                                        "type": "string",
-                                        "description": "Error message"
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
+                ...swaggerErrorNotFound,
             }
         },
         "put": {
@@ -160,22 +155,7 @@ export const expensesSwagger = {
                         }
                     }
                 },
-                "404": {
-                    "description": "Not Found",
-                    "content": {
-                        "application/json": {
-                            "schema": {
-                                "type": "object",
-                                "properties": {
-                                    "error": {
-                                        "type": "string",
-                                        "description": "Error message"
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
+                ...swaggerErrorNotFound,
             }
         },
         "delete": {
@@ -205,85 +185,8 @@ export const expensesSwagger = {
                         }
                     }
                 },
-                "404": {
-                    "description": "Not Found",
-                    "content": {
-                        "application/json": {
-                            "schema": {
-                                "type": "object",
-                                "properties": {
-                                    "error": {
-                                        "type": "string",
-                                        "description": "Error message"
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
+                ...swaggerErrorNotFound,
             }
         }
     }
-}
-
-export const expenseSchemaSwagger = {
-    "Expense": {
-        "type": "object",
-        "required": [
-            "name",
-            "price",
-            "creator"
-        ],
-        "properties": {
-            "id": {
-                "type": "string",
-                "description": "Unique identifier of the expense"
-            },
-            "name": {
-                "type": "string",
-                "description": "Name of the expense"
-            },
-            "price": {
-                "type": "number",
-                "description": "Price of the expense"
-            },
-            "category": {
-                "type": "string",
-                "description": "Expense category (optional)"
-            },
-            "date": {
-                "type": "string",
-                "format": "date",
-                "description": "Date of the expense (optional)"
-            },
-            "creator": {
-                "type": "object",
-                "description": "User who created the expense (limited details)",
-                "properties": {
-                    "id": {
-                        "type": "string",
-                        "description": "User ID"
-                    },
-                    "name": {
-                        "type": "string",
-                        "description": "User name"
-                    },
-                    "image": {
-                        "type": "string",
-                        "description": "User profile image URL (optional)"
-                    }
-                }
-            },
-            "createdAt": {
-                "type": "string",
-                "format": "date-time",
-                "description": "Timestamp of expense creation"
-            },
-            "updatedAt": {
-                "type": "string",
-                "format": "date-time",
-                "description": "Timestamp of expense update"
-            }
-        }
-    },
 }
