+++ src/features/users/users.swagger.api.ts
@@ -1,3 +1,5 @@
+import { swaggerErrorNotFound } from "../../routes/middlewares/swagger/swagger.errors";
+
 export const userSwagger = {
     "/api/users": {
         "get": {
@@ -87,9 +89,7 @@ export const userSwagger = {
                         }
                     }
                 },
-                "404": {
-                    "description": "Not Found"
-                }
+                ...swaggerErrorNotFound,
             }
         },
         "delete": {
@@ -118,47 +118,7 @@ export const userSwagger = {
                         }
                     }
                 },
-                "404": {
-                    "description": "Not Found"
-                }
-            }
-        }
-    },
-}
-
-export const userSchemaSwagger = {
-    "User": {
-        "type": "object",
-        "required": [
-            "name"
-        ],
-        "properties": {
-            "name": {
-                "type": "string",
-                "description": "User's name"
-            },
-            "phone": {
-                "type": "string",
-                "description": "User's phone number (optional)"
-            },
-            "photo": {
-                "type": "string",
-                "description": "URL of the user's photo (optional)"
-            },
-            "password": {
-                "type": "string",
-                "description": "User's password (optional)",
-                "writeOnly": true
-            },
-            "createdAt": {
-                "type": "string",
-                "format": "date-time",
-                "description": "Timestamp of user creation (read-only)"
-            },
-            "updatedAt": {
-                "type": "string",
-                "format": "date-time",
-                "description": "Timestamp of user update (read-only)"
+                ...swaggerErrorNotFound,
             }
         }
     },
