+++ src/routes/middlewares/index.ts
@@ -2,4 +2,4 @@ export * from "./auth0";
 export * from "./catch-all";
 export * from "./error";
 export * from "./validate-resource";
-export * from "./swagger";
+export * from "./swagger/swagger";
