+++ src/routes/middlewares/swagger/swagger.errors.ts
@@ -1,0 +1,18 @@
+export const swaggerErrorNotFound = {
+    "404": {
+        "description": "Not Found",
+        "content": {
+            "application/json": {
+                "schema": {
+                    "type": "object",
+                    "properties": {
+                        "error": {
+                            "type": "string",
+                            "description": "Error message"
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
